+++ server/server-agent-phantomjs.ts
@@ -22,7 +22,7 @@ namespace Server {
 
   // @HINT: Must surround with try/catch otherwise it will block in case of error
   try {
-    baseServerAgent = require('./server-agent.js').Server;
+    baseServerAgent = require('./server-agent.js').ServerAgent;
 
   } catch (error) {
     console.error(error);